tests/submission/PhilipTaylor/canvas/context.emptystring.html

<!DOCTYPE html>
<html><head><meta charset="utf-8">
<title>Canvas test: context.emptystring</title>
<meta content="Philip Taylor" name="author">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../common/canvas-tests.js"></script>
<link href="../common/canvas-tests.css" rel="stylesheet">
</head><body class="show_output">

<h1><a href="index.context.html">context</a>.emptystring</h1>
<p class="desc">getContext with empty string returns null</p>

<div class="refs">Spec references:
<ul>
<li><a href="../annotated-spec/canvas.html#testrefs.context.unrecognised">context.unrecognised</a>
</li></ul>
</div>

<p class="output">Actual output:</p>
<canvas class="output" height="50" id="c" width="100"><p class="fallback">FAIL (fallback content)</p></canvas>

<ul id="d"></ul>
<script>
var t = async_test("getContext with empty string returns null");
_addTest(function(canvas, ctx) {

_assertSame(canvas.getContext(""), null, "canvas.getContext(\"\")", "null");


});
</script>

<div id="log"></div>
</body></html>
